requirements.txt
streamlit>=1.28.0
openai-agents>=0.1.0
playwright>=1.40.0
requests>=2.31.0
beautifulsoup4>=4.12.0
html5lib>=1.1
pandas>=2.0.0
python-dotenv>=1.0.0
